+++ packages/eyes-storybook/test/e2e/happy-config/zero-stories.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  appName: 'Retry when getting zero stories',
+  batchName: 'Retry when getting zero stories',
+  browser: [{width: 800, height: 600, name: 'chrome'}],
+};