+++ packages/eyes-storybook/test/fixtures/zeroStoriesRetry/applitools.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+  concurrency: 1,
+  appName: 'Zero Stories Retry',
+  batchName: 'Zero Stories Retry',
+  storybookConfigDir: 'test/fixtures/zeroStoriesRetry/',
+  storybookStaticDir: 'test/fixtures',
+  storybookPort: 9003,
+};