+++ packages/eyes-storybook/test/fixtures/zeroStoriesRetry/config.js
@@ -1,0 +1,7 @@
+const {configure} = require('@storybook/react');
+
+function loadStories() {
+  require('./index.js');
+}
+
+configure(loadStories, module);