+++ packages/eyes-storybook/test/fixtures/zeroStoriesRetry/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+setTimeout(() => {   
+        storiesOf('zeroStories', module)
+            .add('first story', () => (
+            <div style={{fontSize: '30px'}}>This story came in late</div>
+            ), )
+},1000);