+++ packages/eyes-storybook/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-
+ - add a retry machanisem for cases we get 0 stories
 
 ## 3.25.2 - 2021/10/30